Restructuring the package - first iteration
