+++ pySC/core/constants.py
