+++ pySC/correction/SCfeedback.py
@@ -1,7 +1,7 @@
 import numpy as np
 
-from pySC.core.SCgetBPMreading import SCgetBPMreading
-from pySC.core.SCsetpoints import SCsetCavs2SetPoints, SCsetCMs2SetPoints
+from pySC.core.beam import SCgetBPMreading
+from pySC.core.lattice_setting import SCsetCavs2SetPoints, SCsetCMs2SetPoints
 from pySC.utils import logging_tools
