+++ pySC/lattice_properties/SCgetDispersion.py
@@ -1,7 +1,7 @@
 import numpy as np
 
-from pySC.core.SCgetBPMreading import SCgetBPMreading
-from pySC.core.SCsetpoints import SCsetCavs2SetPoints
+from pySC.core.beam import SCgetBPMreading
+from pySC.core.lattice_setting import SCsetCavs2SetPoints
 
 
 def SCgetDispersion(SC,RFstep,BPMords=None,CAVords=None,nSteps=2):
