+++ pySC/utils/SCmultipolesRead.py
